feat(app): add modal for deleting protocol in ProtocolDashboard (#12825)
Add a confirmation modal when deleting a protocol in ProtocolDashboard. Also move ProtocolDashboard
related components under organisms.
