+++ app/src/App/OnDeviceDisplayApp.tsx
@@ -25,7 +25,7 @@ import { ProtocolSetup } from '../pages/OnDeviceDisplay/ProtocolSetup'
 import { TempODDMenu } from '../pages/OnDeviceDisplay/TempODDMenu'
 import { RobotDashboard } from '../pages/OnDeviceDisplay/RobotDashboard'
 import { RobotSettingsDashboard } from '../pages/OnDeviceDisplay/RobotSettingsDashboard'
-import { ProtocolDashboard } from '../pages/OnDeviceDisplay/ProtocolDashboard'
+import { ProtocolDashboard } from '../pages/ProtocolDashboard'
 import { ProtocolDetails } from '../pages/OnDeviceDisplay/ProtocolDetails'
 import { RunningProtocol } from '../pages/OnDeviceDisplay/RunningProtocol'
 import { RunSummary } from '../pages/OnDeviceDisplay/RunSummary'
