+++ app/src/pages/ProtocolDashboard/__tests__/PinnedProtocol.test.tsx
@@ -3,8 +3,8 @@ import { fireEvent } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
 
 import { renderWithProviders } from '@opentrons/components'
-import { i18n } from '../../../../i18n'
-import { ProtocolCard } from '../'
+import { i18n } from '../../../i18n'
+import { PinnedProtocol } from '../PinnedProtocol'
 
 import type { ProtocolResource } from '@opentrons/shared-data'
 
@@ -21,8 +21,8 @@ jest.mock('react-router-dom', () => {
 const mockProtocol: ProtocolResource = {
   id: 'mockProtocol1',
   createdAt: '2022-05-03T21:36:12.494778+00:00',
-  protocolType: 'json',
   robotType: 'OT-3 Standard',
+  protocolType: 'json',
   metadata: {
     protocolName: 'yay mock protocol',
     author: 'engineering',
@@ -35,12 +35,16 @@ const mockProtocol: ProtocolResource = {
   key: '26ed5a82-502f-4074-8981-57cdda1d066d',
 }
 
-const props = { protocol: mockProtocol, longPress: jest.fn() }
+const props = {
+  protocol: mockProtocol,
+  longPress: jest.fn(),
+  setShowDeleteConfirmationModal: jest.fn(),
+}
 
 const render = () => {
   return renderWithProviders(
     <MemoryRouter>
-      <ProtocolCard {...props} />
+      <PinnedProtocol {...props} />
     </MemoryRouter>,
     {
       i18nInstance: i18n,
@@ -48,7 +52,7 @@ const render = () => {
   )
 }
 
-describe('ProtocolCard', () => {
+describe('Pinned Protocol', () => {
   jest.useFakeTimers()
 
   it('should redirect to protocol details after short click', () => {
@@ -65,6 +69,8 @@ describe('ProtocolCard', () => {
     jest.advanceTimersByTime(1005)
     expect(props.longPress).toHaveBeenCalled()
     getByText('Run protocol')
+    // This should ne "Unpin protocol" but I don't know how to pass state into the render
+    // call so the longpress modal can see the pinned ids.
     getByText('Pin protocol')
     getByText('Delete protocol')
   })
